+++ reprounzip/reprounzip/pack_info.py
@@ -14,6 +14,7 @@ It dispatchs to plugins registered through pkg_resources as entry point
 from __future__ import division, print_function, unicode_literals
 
 import argparse
+import json
 import logging
 import platform
 from rpaths import Path
@@ -26,135 +27,99 @@ from reprounzip.unpackers.common import load_config, COMPAT_OK, COMPAT_MAYBE, \
 from reprounzip.utils import iteritems, itervalues, unicode_, hsize
 
 
-def print_info(args):
-    """Writes out some information about a pack file.
+def get_package_info(pack, read_data=False):
+    """Get information about a package.
     """
-    pack = Path(args.pack[0])
-
-    # Loads config
     runs, packages, other_files = config = load_config(pack)
     inputs_outputs = config.inputs_outputs
 
-    pack_total_size = 0
-    pack_total_paths = 0
-    pack_files = 0
-    pack_dirs = 0
-    pack_symlinks = 0
-    pack_hardlinks = 0
-    pack_others = 0
-    if args.verbosity >= 2:
+    information = {}
+
+    if read_data:
+        total_size = 0
+        total_paths = 0
+        files = 0
+        dirs = 0
+        symlinks = 0
+        hardlinks = 0
+        others = 0
+
         rpz_pack = RPZPack(pack)
         for m in rpz_pack.list_data():
-            pack_total_size += m.size
-            pack_total_paths += 1
+            total_size += m.size
+            total_paths += 1
             if m.isfile():
-                pack_files += 1
+                files += 1
             elif m.isdir():
-                pack_dirs += 1
+                dirs += 1
             elif m.issym():
-                pack_symlinks += 1
+                symlinks += 1
             elif hasattr(m, 'islnk') and m.islnk():
-                pack_hardlinks += 1
+                hardlinks += 1
             else:
-                pack_others += 1
+                others += 1
         rpz_pack.close()
 
-    meta_total_paths = 0
-    meta_packed_packages_files = 0
-    meta_unpacked_packages_files = 0
-    meta_packages = len(packages)
-    meta_packed_packages = 0
+        information['pack'] = {
+            'total_size': total_size,
+            'total_paths': total_paths,
+            'files': files,
+            'dirs': dirs,
+            'symlinks': symlinks,
+            'hardlinks': hardlinks,
+            'others': others,
+        }
+
+    total_paths = 0
+    packed_packages_files = 0
+    unpacked_packages_files = 0
+    packed_packages = 0
     for package in packages:
         nb = len(package.files)
-        meta_total_paths += nb
+        total_paths += nb
         if package.packfiles:
-            meta_packed_packages_files += nb
-            meta_packed_packages += 1
+            packed_packages_files += nb
+            packed_packages += 1
         else:
-            meta_unpacked_packages_files += nb
+            unpacked_packages_files += nb
     nb = len(other_files)
-    meta_total_paths += nb
-    meta_packed_paths = meta_packed_packages_files + nb
+    total_paths += nb
+
+    information['meta'] = {
+        'total_paths': total_paths,
+        'packed_packages_files': packed_packages_files,
+        'unpacked_packages_files': unpacked_packages_files,
+        'packages': len(packages),
+        'packed_packages': packed_packages,
+        'packed_paths': packed_packages_files + nb,
+    }
 
     if runs:
-        meta_architecture = runs[0]['architecture']
-        if any(r['architecture'] != meta_architecture
+        architecture = runs[0]['architecture']
+        if any(r['architecture'] != architecture
                for r in runs):
             logging.warning("Runs have different architectures")
-        meta_distribution = runs[0]['distribution']
-        if any(r['distribution'] != meta_distribution
+        information['meta']['architecture'] = architecture
+        distribution = runs[0]['distribution']
+        if any(r['distribution'] != distribution
                for r in runs):
             logging.warning("Runs have different distributions")
-        meta_distribution = ' '.join(t for t in meta_distribution if t)
+        information['meta']['distribution'] = distribution
 
-    current_architecture = platform.machine().lower()
-    current_distribution = platform.linux_distribution()[0:2]
-    current_distribution = ' '.join(t for t in current_distribution if t)
+        information['runs'] = [
+            dict((k, run[k])
+                 for k in ['id', 'binary', 'argv', 'environ',
+                           'workingdir', 'signal', 'exitcode']
+                 if k in run)
+            for run in runs]
 
-    print("Pack file: %s" % pack)
-    print("\n----- Pack information -----")
-    print("Compressed size: %s" % hsize(pack.size()))
-    if args.verbosity >= 2:
-        print("Unpacked size: %s" % hsize(pack_total_size))
-        print("Total packed paths: %d" % pack_total_paths)
-    if args.verbosity >= 3:
-        print("    Files: %d" % pack_files)
-        print("    Directories: %d" % pack_dirs)
-        if pack_symlinks:
-            print("    Symbolic links: %d" % pack_symlinks)
-        if pack_hardlinks:
-            print("    Hard links: %d" % pack_hardlinks)
-    if pack_others:
-        print("    Unknown (what!?): %d" % pack_others)
-    print("\n----- Metadata -----")
-    if args.verbosity >= 3:
-        print("Total paths: %d" % meta_total_paths)
-        print("Listed packed paths: %d" % meta_packed_paths)
-    if packages:
-        print("Total software packages: %d" % meta_packages)
-        print("Packed software packages: %d" % meta_packed_packages)
-        if args.verbosity >= 3:
-            print("Files from packed software packages: %d" %
-                  meta_packed_packages_files)
-            print("Files from unpacked software packages: %d" %
-                  meta_unpacked_packages_files)
-    if runs:
-        print("Architecture: %s (current: %s)" % (meta_architecture,
-                                                  current_architecture))
-        print("Distribution: %s (current: %s)" % (
-              meta_distribution, current_distribution or "(not Linux)"))
-        print("Runs (%d):" % len(runs))
-        for run in runs:
-            cmdline = ' '.join(shell_escape(a) for a in run['argv'])
-            if len(runs) == 1 and run['id'] == "run0":
-                print("    %s" % cmdline)
-            else:
-                print("    %s: %s" % (run['id'], cmdline))
-            if args.verbosity >= 2:
-                print("        wd: %s" % run['workingdir'])
-                if 'signal' in run:
-                    print("        signal: %d" % run['signal'])
-                else:
-                    print("        exitcode: %d" % run['exitcode'])
-                if run.get('walltime') is not None:
-                    print("        walltime: %s" % run['walltime'])
-
-    if inputs_outputs:
-        if args.verbosity < 2:
-            print("Inputs/outputs files (%d): %s" % (
-                  len(inputs_outputs), ", ".join(sorted(inputs_outputs))))
-        else:
-            print("Inputs/outputs files (%d):" % len(inputs_outputs))
-            for name, f in sorted(iteritems(inputs_outputs)):
-                t = []
-                if f.read_runs:
-                    t.append("in")
-                if f.write_runs:
-                    t.append("out")
-                print("    %s (%s): %s" % (name, ' '.join(t), f.path))
+    information['inputs_outputs'] = {
+        name: {'path': str(iofile.path),
+               'read_runs': iofile.read_runs,
+               'write_runs': iofile.write_runs}
+        for name, iofile in iteritems(inputs_outputs)}
 
     # Unpacker compatibility
-    print("\n----- Unpackers -----")
     unpacker_status = {}
     for name, upk in iteritems(unpackers):
         if 'test_compatibility' in upk:
@@ -168,19 +133,115 @@ def print_info(args):
             unpacker_status.setdefault(compat, []).append((name, msg))
         else:
             unpacker_status.setdefault(None, []).append((name, None))
-    for s, n in [(COMPAT_OK, "Compatible"), (COMPAT_MAYBE, "Unknown"),
-                 (COMPAT_NO, "Incompatible")]:
-        if s != COMPAT_OK and args.verbosity < 2:
-            continue
-        if s not in unpacker_status:
-            continue
-        upks = unpacker_status[s]
-        print("%s (%d):" % (n, len(upks)))
-        for upk_name, msg in upks:
-            if msg is not None:
-                print("    %s (%s)" % (upk_name, msg))
+    information['unpacker_status'] = unpacker_status
+
+    return information
+
+
+def _print_package_info(pack, info, verbosity=1):
+    print("Pack file: %s" % pack)
+    print("\n----- Pack information -----")
+    print("Compressed size: %s" % hsize(pack.size()))
+
+    info_pack = info.get('pack')
+    if info_pack:
+        if 'total_size' in info_pack:
+            print("Unpacked size: %s" % hsize(info_pack['total_size']))
+        if 'total_paths' in info_pack:
+            print("Total packed paths: %d" % info_pack['total_paths'])
+        if verbosity >= 3:
+            print("    Files: %d" % info_pack['files'])
+            print("    Directories: %d" % info_pack['dirs'])
+            if info_pack.get('symlinks'):
+                print("    Symbolic links: %d" % info_pack['symlinks'])
+            if info_pack.get('hardlinks'):
+                print("    Hard links: %d" % info_pack['hardlinks'])
+        if info_pack.get('others'):
+            print("    Unknown (what!?): %d" % info_pack['others'])
+    print("\n----- Metadata -----")
+    info_meta = info['meta']
+    if verbosity >= 3:
+        print("Total paths: %d" % info_meta['total_paths'])
+        print("Listed packed paths: %d" % info_meta['packed_paths'])
+    if info_meta.get('packages'):
+        print("Total software packages: %d" % info_meta['packages'])
+        print("Packed software packages: %d" % info_meta['packed_packages'])
+        if verbosity >= 3:
+            print("Files from packed software packages: %d" %
+                  info_meta['packed_packages_files'])
+            print("Files from unpacked software packages: %d" %
+                  info_meta['unpacked_packages_files'])
+    if 'architecture' in info_meta:
+        print("Architecture: %s (current: %s)" % (info_meta['architecture'],
+                                                  platform.machine().lower()))
+    if 'distribution' in info_meta:
+        distribution = ' '.join(t for t in info_meta['distribution'] if t)
+        current_distribution = platform.linux_distribution()[0:2]
+        current_distribution = ' '.join(t for t in current_distribution if t)
+        print("Distribution: %s (current: %s)" % (
+              distribution, current_distribution or "(not Linux)"))
+    if 'runs' in info:
+        runs = info['runs']
+        print("Runs (%d):" % len(runs))
+        for run in runs:
+            cmdline = ' '.join(shell_escape(a) for a in run['argv'])
+            if len(runs) == 1 and run['id'] == "run0":
+                print("    %s" % cmdline)
             else:
-                print("    %s" % upk_name)
+                print("    %s: %s" % (run['id'], cmdline))
+            if verbosity >= 2:
+                print("        wd: %s" % run['workingdir'])
+                if 'signal' in run:
+                    print("        signal: %d" % run['signal'])
+                else:
+                    print("        exitcode: %d" % run['exitcode'])
+                if run.get('walltime') is not None:
+                    print("        walltime: %s" % run['walltime'])
+
+    inputs_outputs = info.get('inputs_outputs')
+    if inputs_outputs:
+        if verbosity < 2:
+            print("Inputs/outputs files (%d): %s" % (
+                  len(inputs_outputs), ", ".join(sorted(inputs_outputs))))
+        else:
+            print("Inputs/outputs files (%d):" % len(inputs_outputs))
+            for name, f in sorted(iteritems(inputs_outputs)):
+                t = []
+                if f['read_runs']:
+                    t.append("in")
+                if f['write_runs']:
+                    t.append("out")
+                print("    %s (%s): %s" % (name, ' '.join(t), f['path']))
+
+    unpacker_status = info.get('unpacker_status')
+    if unpacker_status:
+        print("\n----- Unpackers -----")
+        for s, n in [(COMPAT_OK, "Compatible"), (COMPAT_MAYBE, "Unknown"),
+                     (COMPAT_NO, "Incompatible")]:
+            if s != COMPAT_OK and verbosity < 2:
+                continue
+            if s not in unpacker_status:
+                continue
+            upks = unpacker_status[s]
+            print("%s (%d):" % (n, len(upks)))
+            for upk_name, msg in upks:
+                if msg is not None:
+                    print("    %s (%s)" % (upk_name, msg))
+                else:
+                    print("    %s" % upk_name)
+
+
+def print_info(args):
+    """Writes out some information about a pack file.
+    """
+    pack = Path(args.pack[0])
+
+    info = get_package_info(pack, read_data=args.json or args.verbosity >= 2)
+    if args.json:
+        json.dump(info, sys.stdout, indent=2)
+        sys.stdout.write('\n')
+    else:
+        _print_package_info(pack, info, args.verbosity)
 
 
 def showfiles(args):
@@ -311,6 +372,7 @@ def setup_info(parser, **kwargs):
     """
     parser.add_argument('pack', nargs=1,
                         help="Pack to read")
+    parser.add_argument('--json', action='store_true', default=False)
     parser.set_defaults(func=print_info)