+++ reprozip/reprozip/tracer/trace.py
@@ -15,6 +15,7 @@ from collections import defaultdict
 from itertools import count
 import logging
 import os
+from pkg_resources import iter_entry_points
 import platform
 from rpaths import Path
 import sqlite3
@@ -88,6 +89,15 @@ class TracedFile(File):
                 self.runs[run] = TracedFile.WRITTEN
             elif self.runs[run] == TracedFile.ONLY_READ:
                 self.runs[run] = TracedFile.READ_THEN_WRITTEN
+
+
+def run_filter_plugins(files, input_files):
+    for entry_point in iter_entry_points('reprozip.filters'):
+        func = entry_point.load()
+        name = entry_point.name
+
+        logging.info("Running filter plugin %s", name)
+        func(files=files, input_files=input_files)
 
 
 def get_files(conn):
@@ -203,6 +213,13 @@ def get_files(conn):
                         for m in magic_dirs + system_dirs)]
                for r, lst in enumerate(access_files)]
 
+    # Run the list of files through the filter plugins
+    run_filter_plugins(files, inputs)
+
+    # Files removed from plugins should be removed from inputs as well
+    inputs = [[path for path in lst if path in files]
+              for lst in inputs]
+
     # Displays a warning for READ_THEN_WRITTEN files
     read_then_written_files = [
         fi
@@ -266,6 +283,7 @@ def tty_prompt(prompt, chars):
             while True:
                 char = stream.read(1)
                 if char in chars:
+                    stream.write("\n")
                     return char
         finally:
             termios.tcsetattr(fd, tcsetattr_flags, old)